Move templated code to its own header
