+++ mlir/include/mlir/Dialect/Rock/IR/GetRockInfo.h
@@ -53,62 +53,6 @@ rock::GemmFeatures getFeatures(Operation *op);
 // an optional 'GemmFeatures' attribute
 bool opHasOptionalFeature(Operation *op);
 
-// This function returns the func or gpu.func of a given op
-inline Operation *getParentFuncOp(Operation *op) {
-  Operation *func;
-  if (isa<func::FuncOp, gpu::GPUFuncOp>(op)) {
-    func = op;
-  } else {
-    func = op->getParentOfType<func::FuncOp>();
-    if (!func) {
-      func = op->getParentOfType<gpu::GPUFuncOp>();
-    }
-  }
-
-  return func;
-}
-
-// Helper function to get attributes from parents
-template <typename RetAttrType>
-FailureOr<RetAttrType> getAttrFromOpOrParents(
-    Operation *op, StringRef opAttr,
-    std::optional<StringRef> maybeDialectAttr = std::nullopt) {
-  StringRef dialectAttr = maybeDialectAttr.value_or(opAttr);
-  Operation *func = getParentFuncOp(op);
-  RetAttrType attr;
-  auto getAnyAttr = [&](ArrayRef<StringRef> attrNames, Operation *op) {
-    for (StringRef attrName : attrNames) {
-      if (!attr) {
-        attr = op->getAttrOfType<RetAttrType>(attrName);
-      } else {
-        return;
-      }
-    }
-  };
-
-  // First check for the attribute on the op
-  getAnyAttr({opAttr}, op);
-  if (!attr) {
-    // If that fails then try checking for the attribute on the func
-    getAnyAttr({opAttr, dialectAttr}, func);
-  }
-
-  // If there is no desired attribute on the func, then check the nearest parent
-  // with a symbol table (covers both ModuleOp and gpu::GPUModuleOp)
-  if (!attr) {
-    if (auto symbolTableOp = func->getParentWithTrait<OpTrait::SymbolTable>()) {
-      getAnyAttr({opAttr, dialectAttr}, symbolTableOp);
-      if (attr)
-        return attr;
-    }
-  }
-
-  if (!attr) {
-    return failure();
-  }
-  return attr;
-}
-
 } // End namespace rock
 } // End namespace mlir
 #endif // MLIR_DIALECT_ROCK_IR_GETROCKINFO_H
